Code split - fase 1 Create InputArea
